Default to UK elections app in all test
This change is made as part of removing other countries from the
UK fork of YNR. The change is simple, but the result is that a lot of
tests fail because of some UK specific template and data model changes.

Because we're only concerned with testing the UK app, the tests have
been updated to expect the expected UK values.